+++ contracts/Core/interface/ICollectionManager.sol
@@ -67,7 +67,7 @@ interface ICollectionManager {
      * @return result of the collection
      * @return power of the resultant collection
      */
-    function getResult(bytes32 _name) external view returns (uint32, int8);
+    function getResult(bytes32 _name) external view returns (uint256, int8);
 
     /**
      * @notice returns the result of the collection based on the id sent by the client
@@ -75,7 +75,7 @@ interface ICollectionManager {
      * @return result of the collection
      * @return power of the resultant collection
      */
-    function getResultFromID(uint16 _id) external view returns (uint32, int8);
+    function getResultFromID(uint16 _id) external view returns (uint256, int8);
 
     /**
      * @return epoch in which the registry needs to be updated
